backend/requirements.txt
flask
flask-cors
python-dotenv
scikit-learn
pandas
numpy
joblib
requests
groq
